requirements.txt
aws-cdk.core

aws-cdk.aws-autoscaling
aws-cdk.aws-ec2
aws-cdk.aws-elasticloadbalancingv2
aws-cdk.aws-certificatemanager
aws-cdk.aws-route53
aws-cdk.aws-route53-targets

# Work around for jsii#413
aws-cdk.aws-autoscaling-common
